refactor: Split join search module into focused submodules
Split the monolithic 919-line mod.rs into smaller, focused files:

- config.rs: ParallelSearchConfig and environment variable parsing
- context.rs: JoinOrderContext struct definition
- state.rs: SearchState and JoinCost types
- optimizer.rs: JoinOrderSearch public API
- tests.rs: Comprehensive test suite (260+ lines)

Benefits:
- Improved readability: Each file has a single, clear responsibility
- Easier navigation: Developers can find specific functionality quickly
- Better testability: Logic is isolated and can be tested independently
- Reduced cognitive load: Files are ~50-200 lines instead of 900+

No functional changes to the optimizer. All existing tests pass.

Closes #2484
