+++ crates/vibesql-executor/src/select/join/search/context.rs
@@ -1,0 +1,25 @@
+//! Context for join order search operations
+
+use std::collections::BTreeSet;
+
+use super::config::ParallelSearchConfig;
+use super::reorder::JoinEdge;
+
+/// Context for join order search operations
+///
+/// Holds shared state and configuration used by all search strategies.
+/// This internal struct encapsulates data needed by cost estimation,
+/// DFS, BFS, and greedy algorithms.
+pub(super) struct JoinOrderContext {
+    /// All tables in the query
+    pub all_tables: BTreeSet<String>,
+    /// Join edges (which tables connect)
+    pub edges: Vec<JoinEdge>,
+    /// Estimated rows for each table after local filters
+    pub table_cardinalities: std::collections::HashMap<String, usize>,
+    /// Selectivity for each join edge based on column NDV (number of distinct values)
+    /// Key is (left_table, right_table) normalized to lowercase
+    pub edge_selectivities: std::collections::HashMap<(String, String), f64>,
+    /// Configuration for parallel search
+    pub config: ParallelSearchConfig,
+}
